Update recover.c
I liked the resolution a lot! It helped me to solve the problem.

I realized there's no need to copy and paste the part of the code that creates a new file if the counter is 0 or don't. Just use an if condition if it's not zero and close the previous file. It makes the resolution more succinct and saves steps.

I also added a constant BLOCK_SIZE and a return value for errors and success.